+++ CHANGELOG.md
@@ -73,3 +73,11 @@
 - Added my own meter element for a better UI
 - Now the weight and height of the pokémon is made in HTML
 - UI improvements in the stats page
+
+## 1.4.0
+
+### Descriptions and navigation
+
+- Changed the clear button
+- Added navigation to the stats page, with a back button and the pokemon ID
+- Added descriptions to the stats page